[FEATURE] Add flush preview image button
It is now possible to selectively delete Youtube preview images from
within the backend. A new button is added next to the inline record
control buttons (right next to the drag move button). A notification
will appear upon clicking, which will notify the user of the success of
the operation.